Add conda configuration specifying matrix of python and gnuradio versions
This will cause conda-build to create a build for each supported version
of python to match the version of python that the corresponding gnuradio
package was built for.

In order to get the python build variants to appear in the build string,
the git hash (for non-release-tagged builds) has been moved to be
a part of the version identifier, not the build string.