conda: Depend on same version of compilers as gnuradio
The gnuradio conda packaging at times pins to older compiler versions
than conda-forge/conda-build will provide by default. This causes all
sorts of fun specifically with pybind11, which is a header-only C++
library containing global variables that need to be shared between
gnuradio-core and the OOT modules (like this one) that use it.